services/pricing-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.0
psycopg2-binary==2.9.9